query-sandbox-lib/src/main/java/com/github/barrettotte/querysandbox/service/MetricService.java
package com.github.barrettotte.querysandbox.service;

import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.stereotype.Service;

@Service
@EnableConfigurationProperties(ServiceProperties.class)
public class MetricService {

    private final ServiceProperties serviceProperties;

    public MetricService(ServiceProperties serviceProperties) {
        this.serviceProperties = serviceProperties;
    }

    public String message() {
        return this.serviceProperties.getMessage();
    }
}
